requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
httpx==0.28.1
pandas==3.0.1
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.8.1
google-genai==1.5.0
